// src/app/interface/product.interface.ts
export interface IProduct {
  title: string;
  description: string;
  price: number;
  discountPercentage: number;
  rating: number;
  stock: number;
  available: boolean;
  brand: string;
  category: string;
  thumbnail: string;
}

// {
//   "_id": {
//     "$oid": "67be03249ff0b0413e7a8155"
//   },
//   "title": "iPhone 9",
//   "description": "An apple mobile which is nothing like apple",
//   "price": 993,
//   "discountPercentage": 12.96,
//   "rating": 0,
//   "stock": 0,
//   "brand": "Apple",
//   "category": "smartphones",
//   "thumbnail": "https://i0.wp.com/9to5mac.com/wp-content/uploads/sites/6/2019/10/iphone-8-1.png?resize=1200%2C628&quality=82&strip=all&ssl=1",
//   "images": [
//     "https://i.dummyjson.com/data/products/1/1.jpg",
//     "https://i.dummyjson.com/data/products/1/2.jpg",
//     "https://i.dummyjson.com/data/products/1/3.jpg",
//     "https://i.dummyjson.com/data/products/1/thumbnail.jpg"
//   ],
//   "deleted": true
// }